% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_history.R
\name{check_history}
\alias{check_history}
\title{Check validity of history data}
\usage{
check_history(history)
}
\arguments{
\item{history}{a data.frame, containing historic data for one or several
variables}
}
\description{
Modifies order, classes, if necessary.
}
\examples{

check_history(swiss_history)

}
