% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono_strat.R
\name{chrono_bldr}
\alias{chrono_bldr}
\title{Chronostratigraphic columns for ggplot}
\usage{
chrono_bldr(graph, reverse = FALSE, capture_legend = FALSE)
}
\arguments{
\item{graph}{ggplot object}
}
\value{
ggplot
}
\description{
Chronostratigraphic columns for ggplot
}
\examples{
library(tidyverse)
library(pangaear)
#  Westherold et al 2020 ref curve on Pangaea database
Ceno_doi <- pg_search('Cenozoic global reference benthic carbon and oxygen isotope dataset (CENOGRID)')
CD_iso_dat <- pg_data(Ceno_doi$doi)
ref_curve <- CD_iso_dat[[24]]$data

gg_iso <- ggplot(ref_curve,
  aes(x = `Tuned time [Ma]`,
      y = `Foram bent δ18O [‰ PDB] (VPDB)`,
      color = `Reference (benthic Source)`
      )
      ) +
  geom_point() +
  geom_line () +
  scale_x_continuous(trans = "reverse", expand = c(0,0)) +
  coord_cartesian(xlim = c(55, 50), expand=FALSE)

gg_lst <- chrono_bldr(gg_iso, reverse = TRUE)
}
