% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono_chart.R
\docType{data}
\name{chrono_chart}
\alias{chrono_chart}
\title{The chronostratigraphic timeascles with colors}
\format{
A data frame with 36 rows and 7 variables:
\describe{
\item{name}{name of interval}
\item{type}{hierarchy of stratigraphic units}
\item{top}{upper bound in Myr}
\item{bottom}{lower bound in Myr}
\item{R}{R}
\item{G}{G}
\item{B}{B}
}
}
\source{
\url{https://stratigraphy.org/chart}
}
\usage{
chrono_chart
}
\description{
A dataset containing the  stratigraphic timescale with color in RGB
}
\keyword{datasets}
