% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_transients.R
\docType{data}
\name{clim_transients}
\alias{clim_transients}
\title{Climatic events (transients)}
\format{
A data frame with 6 rows and 4 variables:
\describe{
\item{label}{name of the climatic events}
\item{label_exp}{extended description of the climatic events}
\item{x}{mid point in Myr}
\item{y}{based on nearest point in Proxy variable of temp_curve}
}
}
\source{
various sources
}
\usage{
clim_transients
}
\description{
A dataset containing the dramatic hyperthermals and ice age transitions
}
\keyword{datasets}
