% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clim_trends.R
\docType{data}
\name{clim_trends}
\alias{clim_trends}
\title{Climate trends}
\format{
A data frame with 3 rows and 5 variables:
\describe{
\item{xmin}{lower bound in Myr}
\item{xmax}{upper bound in Myr}
\item{x}{mid point in Myr}
\item{y}{based on nearest point in Proxy variable of temp_curve}
\item{label}{nameof the climatic trend}
}
}
\source{
various sources
}
\usage{
clim_trends
}
\description{
A dataset containing the ranges of certain climatic trends
}
\keyword{datasets}
