% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cristal_ball.R
\name{cristal_ball}
\alias{cristal_ball}
\title{GUI for timemachine history data}
\usage{
cristal_ball(history)
}
\arguments{
\item{history}{a data frame with the publication history of the data.
Must have column names \code{pub_date}, \code{ref_date} and \code{value}, plus a column
that identifies multiple series. Use \code{check_history()} to check the validity
of a history data.frame}
}
\description{
GUI for timemachine history data
}
\examples{
\dontrun{
cristal_ball(swiss_history)
}
}
