% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climchange_model.R
\name{curve_fit}
\alias{curve_fit}
\title{Fit growth model to climate curve}
\usage{
curve_fit(df, time, proxy)
}
\arguments{
\item{df}{data for curve fitting exercise}

\item{time}{variable of time units}

\item{proxy}{variable of the proxy}
}
\value{
list of elements
}
\description{
Fit growth model to climate curve
}
