% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antarctica.R
\name{gg_ACC}
\alias{gg_ACC}
\title{Create map of Antarctica with Antarctic Circumpolar Current}
\usage{
gg_ACC(event, ...)
}
\arguments{
\item{event}{character string of event}

\item{...}{not supported

returned}
}
\value{
ggplot
}
\description{
Create map of Antarctica with Antarctic Circumpolar Current
}
