% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panama.R
\name{gg_GS}
\alias{gg_GS}
\title{Create map of Panama and the influence on sea currents}
\usage{
gg_GS(event, ...)
}
\arguments{
\item{event}{character string of event}

\item{...}{not supported

returned}
}
\value{
ggplot
}
\description{
Create map of Panama and the influence on sea currents
}
