% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/india.R
\name{gg_HIM}
\alias{gg_HIM}
\title{Create map of India depicting the collision with Eurasia}
\usage{
gg_HIM(event, ...)
}
\arguments{
\item{event}{character string of event}

\item{...}{not supported

returned}
}
\value{
ggplot
}
\description{
Create map of India depicting the collision with Eurasia
}
