% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tectodriver.R
\name{gg_tect}
\alias{gg_tect}
\title{Create range plot of tectonic drivers}
\usage{
gg_tect(lims, ...)
}
\arguments{
\item{lims}{numeric vector of length 2 giving the limits of the timeplot}

\item{...}{not supported

returned}
}
\value{
ggplot
}
\description{
Create range plot of tectonic drivers
}
