% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{latest}
\alias{latest}
\title{Latest available data}
\usage{
latest(history)
}
\arguments{
\item{history}{a data.frame, containing historic data for one or several
variables}
}
\description{
Latest available data
}
