\name{pdm}
\docType{data}
\alias{pdm}
\title{Example Dataset}
\description{
Dataset describing an initial population of 100 individuals subdivided into 1024
types, corresponding to each possible combination of genotypes over 10 loci.
}
\usage{pdm}
\format{\describe{
    \item{unitary.transitions}{Per-site transition matrix}
    \item{loci}{Number of loci}
    \item{population}{Population}
}}
\keyword{datasets}

