% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_history.R
\name{pseudo_history}
\alias{pseudo_history}
\title{Pseudo History Data Frame}
\usage{
pseudo_history(x, by = NULL)
}
\arguments{
\item{x}{a \code{ts_boxable} time series.}

\item{by}{offset of publication date. See details.}
}
\description{
If no historical data is available, \emph{pseudo history} data can be produced
from current data. The function replicates the data for each observational
point. If \code{by} is specified, a publication delay will be assumed.
}
\details{
\code{by} can be specified in several ways.
\itemize{
\item A number, taken to be in days.
\item A object of class \code{difftime}
\item A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".
}
}
\examples{
library(tsbox)
pseudo_history(ts_c(mdeaths, fdeaths), "1 month")
pseudo_history(fdeaths)
}
