% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_instrum_data.R
\name{read_cru_hemi}
\alias{read_cru_hemi}
\title{The read_cru_hemi function on http://www.cru.uea.ac.uk/cru/data/temperature}
\usage{
read_cru_hemi(filename)
}
\arguments{
\item{filename}{file from the website}
}
\value{
dataframe
}
\description{
The read_cru_hemi function on http://www.cru.uea.ac.uk/cru/data/temperature
}
\details{
Reads a CRU-format hemispheric average file, as provided at
http://www.cru.uea.ac.uk/cru/data/temperature

Format has two lines for each year.

Returns a data frame with columns:
year (1850 to final year)
annual (mean annual anomaly)
month.1 ... month.12 (mean monthly anomaly)
cover.1 ... cover.12 (percentage coverage)
}
