\name{stationary.dist}
\alias{stationary.dist}
\title{Compute Stationary Distribution of a Markov Chain}
\description{Computes the stationary distribution of a Markov chain specified by
             its transition matrix.}
\usage{stationary.dist(transitions)}
\arguments{
    \item{transitions}{Square matrix specifying the transition probabilities of
                       the Markov chain}
}
\value{A vector representing the stationary distribution.}

