% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tect_events.R
\docType{data}
\name{tect_events}
\alias{tect_events}
\title{Tectonic events}
\format{
A data frame with 3 rows and 6 variables:
\describe{
\item{label}{nameof  the tectonic event}
\item{y}{numerical order}
\item{xmin}{lower bound in Myr}
\item{xmax}{upper bound in Myr}
\item{ref}{Reference}
\item{xrange}{range in Myr}
\item{xmid}{mid value in Myr}
}
}
\source{
various sources
}
\usage{
tect_events
}
\description{
A dataset containing the ranges of Cenozoic tectonic events
}
\keyword{datasets}
