% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeplot.R
\name{time_plot}
\alias{time_plot}
\title{Create proxy timeseries ggplot}
\usage{
time_plot(
  df,
  time,
  proxy,
  events = TRUE,
  explain = FALSE,
  range_sh = NULL,
  ice = FALSE
)
}
\arguments{
\item{df}{data}

\item{time}{variable with time units}

\item{proxy}{variable with proxy data}

\item{events}{logical whether to include climate event labels}

\item{explain}{logical whether the labels include explanatory data}

\item{range_sh}{the range of the window selected}
}
\value{
ggplot
}
\description{
Create proxy timeseries ggplot
}
