% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{timemachine}
\alias{timemachine}
\title{Travel Through Time}
\usage{
timemachine(..., timemachine.dates = getOption("timemachine.dates"))
}
\description{
Travel Through Time
}
\examples{
library(tsbox)
library(timemachine)
library(dplyr)
library(tidyr)

# Constructing pseudo history data. It is assumed that mdeath is available 
# one month after the end of the period, but fdeath immediately. 
timemachine.history <- bind_rows(
  pseudo_history(ts_tbl(mdeaths), "1 month"),
  pseudo_history(ts_tbl(fdeaths))
)

options(timemachine.history = timemachine.history)

# options
options(timemachine.expose = c("ts", "data.frame"))
options(timemachine.expose = c("ts", "tbl"))
options(timemachine.dates = seq(as.Date("1979-10-01"), to = as.Date("1979-12-01"), by = "month"))

library(forecast)
timemachine({
  m <- forecast(auto.arima(mdeaths))
  m$mean   # expression must evaluate to a tsboxable object
})

# Using multiple experessions
timemachine(
  etf = {
    m <- forecast(mdeaths)
    m$mean 
  },
  arima = {
    m <- forecast(auto.arima(mdeaths))
    m$mean
  }
)

}
