% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine_app.R
\name{timemachine_app}
\alias{timemachine_app}
\title{The time machine app}
\usage{
timemachine_app(...)
}
\arguments{
\item{...}{no entries suported here}
}
\value{
shiny app
}
\description{
The time machine app
}
