% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine_app.R
\docType{data}
\name{transients}
\alias{transients}
\title{#' Run the Shiny Application
#'
#' @param ... A series of options to be used inside the app.
#'
#' @export
#' @importFrom golem with_golem_options
run_app <- function(...) {
with_golem_options(
app = shinyApp(ui = ui, server = server),
golem_opts = list(...)
)
}}
\format{
An object of class \code{character} of length 3.
}
\usage{
transients
}
\description{
#' Run the Shiny Application
#'
#' @param ... A series of options to be used inside the app.
#'
#' @export
#' @importFrom golem with_golem_options
run_app <- function(...) {
with_golem_options(
app = shinyApp(ui = ui, server = server),
golem_opts = list(...)
)
}
}
\keyword{datasets}
