% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_to_date}
\alias{add_to_date}
\title{Add Days, Weeks, Months, Quarters or Years to Dates}
\usage{
add_to_date(x, by)
}
\arguments{
\item{by}{can be specified in several ways.
- A number, taken to be in days.
- A object of class difftime
- A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".}
}
\description{
Add Days, Weeks, Months, Quarters or Years to Dates
}
