% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrono_strat.R
\name{chrono_bldr}
\alias{chrono_bldr}
\title{Chronostratigraphic columns for ggplot}
\usage{
chrono_bldr(graph, reverse = FALSE, capture_legend = FALSE, tectonic = FALSE)
}
\arguments{
\item{graph}{ggplot object}

\item{reverse}{logical whether the time axis was already reversed}

\item{capture_legend}{logical indicating whether the legend should be
returned}

\item{tectonic}{logical indicating whether tectonic event plot is required}
}
\value{
A list with ggplot, gtable and grob elements
}
\description{
Chronostratigraphic columns for ggplot
}
