% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_date.R
\name{parse_date_2000colon1}
\alias{parse_date_2000colon1}
\title{Convert to date}
\usage{
parse_date_2000colon1(x)
}
\arguments{
\item{character}{vector}
}
\description{
Date converter with similar stucture
- Naming convention: `parse_date_[EXAMPLE]
- Apply operation on unique values only (use `apply_to_unique()`)
- Use `stringi`?
}
\examples{
parse_date_2000colon1(c("2000:1", "2000:2"))

# These should drop an error:
# parse_date_2000colon1(c("2000:5"))
# parse_date_2000colon1(c("00:1"))
}
