% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pseudo_history}
\alias{pseudo_history}
\title{Pseudo History Data Frame}
\usage{
pseudo_history(x, by = NULL)
}
\description{
Pseudo History Data Frame
}
\examples{
pseudo_history(ts_tbl(mdeaths), "1 month")
pseudo_history(ts_tbl(fdeaths))
}
