% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{timemachine}
\alias{timemachine}
\title{Travel Through Time}
\usage{
timemachine(..., dates = getOption("timemachine.dates"))
}
\arguments{
\item{...}{expressions to be evaluated. Expressions can be named, so the name
will appear in the output}

\item{dates}{expressions to be evaluated. Usually specifyied via
options. See examples. Expressions can be named, so the name
will appear in the output}
}
\description{
Core function of the package. Runs through a series of dates and evaluates
one or several expressions with the data available. Collects `ts_boxable`
objects returned by the expressions.
}
\examples{
library(tsbox)
library(timemachine)
library(dplyr)
library(tidyr)

# Constructing pseudo history data. It is assumed that mdeath is available
# one month after the end of the period, but fdeath immediately.
ldeaths.history <- bind_rows(
  pseudo_history(ts_tbl(mdeaths), "1 month"),
  pseudo_history(ts_tbl(fdeaths))
)

options(timemachine.history = ldeaths.history)
options(dates = seq(as.Date("1979-10-01"), to = as.Date("1979-12-01"), by = "month"))

library(forecast)
timemachine({
  m <- forecast(auto.arima(mdeaths))
  m$mean   # expression must evaluate to a tsboxable object
})

# Using multiple experessions
timemachine(
  etf = {
    m <- forecast(mdeaths)
    m$mean
  },
  arima = {
    m <- forecast(auto.arima(mdeaths))
    m$mean
  }
)

}
