% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{wormhole}
\alias{wormhole}
\title{Travel to a Specific Point in Time}
\usage{
wormhole(
  history,
  date = NULL,
  envir = globalenv(),
  verbose = TRUE,
  post_process = ts_attach
)
}
\arguments{
\item{history}{a data frame with the publication history of the data.
Must have column names \code{pub_date}, \code{ref_date} and \code{value}, plus a column
that identifies multiple series. Use \code{check_history()} to check the validity
of a history data.frame}

\item{date}{\code{"Date"} or \code{"character"}, the date where to a wormhole should be
opened.}

\item{envir}{environment where to expose the data.}

\item{verbose}{logical, what objects will be exposed?}

\item{post_process}{a function that is applied on \code{DATA}. By default,
\code{ts_attach} adds makes series available as single \code{ts} objects. Alternatives
are: \code{ts_ts} (to convert it to a multiple \code{mts} object), or any converter
from tsbox.}
}
\description{
Opening a wormhole enables time travel to a specific point in time. By
default, the \code{history} data frame is exposed to the global
environment as it was available on that day.
}
