% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{wormhole}
\alias{wormhole}
\title{Travel to a Specific Point in Time}
\usage{
wormhole(date = NULL, history = getOption("timemachine.history"),
  envir = timemachinenv(), verbose = TRUE)
}
\arguments{
\item{date}{`"Date"` or `"character"`, the date where to a wormhole should be
opened.}

\item{history}{a data frame containing the history of the data.
Must have column names `pub_date`, `ref_date` and `value`, plus a column
that idenfies multiple series. Currently, this must be a single column
`id`, but this will change. Usually, `history` is set as an
option (see examples)}

\item{envir}{environment where to expose the data.}

\item{verbose}{logical, what objects will be exposed?}

\item{as}{character, how to expose the variables in history. By default, the
function exposes time series as single `"ts"` objects and a data frame,
named `.data`, but all `ts_boxable` objects are supported.}
}
\description{
Opening a wormhole enables time travel to a specific point in time. By
default, the `history` data frame is exposed to the global
enviroment as it was available on that day.
}
