% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wormhole.R
\name{wormhole}
\alias{wormhole}
\title{Travel to a point in Time}
\usage{
wormhole(date = NULL,
  timemachine.history = getOption("timemachine.history"),
  timemachine.expose = getOption("timemachine.expose", c("ts", "data.frame")),
  envir = globalenv(), verbose = TRUE)
}
\description{
Travel to a point in Time
}
